Introduce base class with variable horizon check; integrate with AdversarialTrainer and PreferenceComparison
